requirements.txt
torch>=2.3.0
torchvision>=0.18.0
diffusers>=0.31.0
transformers>=4.46.0
accelerate>=0.33.0
sentencepiece>=0.2.0
ftfy>=6.1.0
gradio>=4.44.0
imageio[ffmpeg]>=2.34.0
opencv-python>=4.10.0.84
numpy>=1.26.0
Pillow>=10.3.0
bitsandbytes